chore: Set up database in development environment with SQLite and Prisma
This is because SQLite does not require installation as it saves the
database in physical files.
Prisma allows working with operations in the database using JS code.
Some advantages are that if we later change the application's database,
it will not  be necessary to rewrite the code for a new language as it
does it automatically; and also supports a wide variety of databases (
PostgreSQL, MySQL, SQLite, MongoDB...).