feat: Add Axios library to integrate the project's frontend and backend
Axios is a promise based HTTP client for the browser and Node.js. It was
used in this project to make HTTP requests from node.js and can also be
used to make XMLHttpRequests from the browser.

For the requests to work, it was also necessary to add CORS to the
project, which is an HTTP-header based mechanism that allows the server
to indicate any origins other than its own from which a browser or the
client side should permit loading resources.